% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execHEG.R
\name{execHEG}
\alias{execHEG}
\title{call the command of the HEG TOOL}
\usage{
execHEG(type = "GRID", infiles, objectname = NULL, fieldname = NULL,
  subset.up = 90, subset.left = -180, subset.down = -90,
  subset.right = 180, resample = "NN", ignore.stderr = NULL,
  Sys_ignore.stdout = FALSE, Sys_wait = TRUE, Sys_input = NULL,
  intern = NULL)
}
\arguments{
\item{type}{a character of "GRID" or "SWATH", specifying the object type of HEG-EOS files}

\item{infiles}{a character vector of HEG-EOS files}

\item{objectname}{character, the SWATH name or GRID name}

\item{fieldname}{character, one filed in the SWATH object}

\item{subset.up}{numeric, the up latitude of the subsetting extent}

\item{resample}{character, the resampling method}
}
\value{
a status code (0 success, -1 failed)
}
\description{
call the command of the HEG TOOL
}
\details{
see the HEG website \url{https://newsroom.gsfc.nasa.gov/sdptoolkit/HEG/HEGHome.html}
}
\note{
set the subsetting extent to the whole geographic region (-180, 180 and -90, 90) when the subsetting operation should not be done
}
\examples{
HEGHOME <- "usr/local/heg"
initHEG(HEGHOME)
objname <- "mod04"
fieldname <- "Optical_Depth_Land_And_Ocean"
infiles = paste0("data_swath/",list.files("data_swath/"))
execHEG(type = "SWATH", infiles = infiles,
       objectname = objname, fieldname = fieldname, Sys_ignore.stdout = FALSE)

}
