% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initHEG.R
\name{initHEG}
\alias{initHEG}
\title{initialize the configuration of HEG}
\usage{
initHEG(HEGHOME)
}
\arguments{
\item{HEGHOME}{the path to the HEG home directory}
}
\description{
initialize the configuration of HEG
}
\note{
the subdirectorIES of HEGHOME must include bin, data, and TOOLKIT_MTD
}
\seealso{
\link{execHEG}
}
